requirements.txt
numpy==1.16.2
pandas==0.23.4
dash_table==3.6.0
redis==2.10.6
dash_html_components==0.14.0
dash_core_components==0.44.0
dash==0.39.0
PeakUtils==1.3.2
dash_callback_chain==0.0.2
plotly==2.7.0
scikit_learn==0.20.3
python_twitter==3.5
xgboost==0.82
gspread==3.1.0
oauth2client==4.1.3
praw==6.1.1
dash_bootstrap_components==0.3.5
